+++ src/Loader/index.js
@@ -51,6 +51,7 @@ const loader = function (content, map, meta) {
     resolve(result != null ? result : content);
   })
     .then((value) => {
+      //return value;
       errorStage = 'compile';
       return Template.compile(value, resource);
     })